Article_API/requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudinary==1.29.0
Deprecated==1.2.13
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.1
fastapi==0.73.0
fastapi-mail==1.0.4
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
packaging==21.3
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
python-decouple==3.6
python-jose==3.3.0
python-multipart==0.0.5
redis==4.1.4
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
wrapt==1.13.3
